pyproject.toml
[project]
name = "linkedin-agent"
version = "0.1.0"
description = "LinkedIn post generator using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community",
    "langchain-core",
    "langgraph",
    "pydantic",
    "ollama",
    "python-dotenv",
    "fastapi",
    "uvicorn"
]

[tool.setuptools]
package-dir = {"" = "src"}
